code refactor for user directory (login + register function -> user)
