+++ food-ordering/app/app/urls.py
@@ -25,7 +25,7 @@ from django.conf.urls.static import static
 urlpatterns = [
     path('', include('starting.urls')),
     path('', include('django.contrib.auth.urls')),
-    # path('user', include('user.urls')),
+    path('user/', include('user.urls')),
     path('admin/', admin.site.urls),
 ]
