+++ food-ordering/app/productionfiles/js/cart.js
@@ -1,0 +1,11 @@
+let iconCart = document.querySelector('.icon-cart');
+let closeBtn = document.querySelector('.cartTab .closeBtn');
+let body = document.querySelector('body');
+
+iconCart.addEventListener('click', () => {
+    body.classList.toggle('activeTabCart');
+})
+
+closeBtn.addEventListener('click', () => {
+    body.classList.toggle('activeTabCart');
+})
