+++ food-ordering/app/starting/forms.py
@@ -1,26 +1,9 @@
-from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 from django.contrib.auth.models import User
 from django import forms
 from django.forms import ModelForm
-from django.forms.widgets import PasswordInput, TextInput
+from django.forms.widgets import TextInput
 from . models import Reservation
 
-
-# - Create/Register a user
-
-class RegisterForm(UserCreationForm):
-
-    class Meta:
-        model = User
-        fields = ['username', 'email', 'password1', 'password2']
-
-# - Authentication Form
-
-class LoginForm(AuthenticationForm):
-    username = forms.CharField(label="Username", widget=TextInput())
-    password = forms.CharField(label="Password", widget=PasswordInput())
-    
-    
 # Reservation Form
 
 VIP_CHOICES = (
