+++ food-ordering/app/starting/templates/base_layout/navbar.html
@@ -55,7 +55,7 @@
                            type="submit"
                            class="btn btn-primary py-2 ms-2 px-4">Log out</a>
                     {% else %}
-                        <a href="login" class="btn btn-primary py-2 ms-2 px-4">Log in</a>
+                        <a href="{% url 'login' %}" class="btn btn-primary py-2 ms-2 px-4">Log in</a>
                     {% endif %}
                 </div>
             </nav>
