+++ food-ordering/app/user/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
+from .models import User1
 
 # Register your models here.
+
+admin.site.register(User1)
