+++ food-ordering/app/user/templates/user/login.html
@@ -100,7 +100,7 @@
                             <div class="row mb-4 justify-content-between">
                                 <!-- Simple link -->
                                 <p>
-                                    <a href="register">Create new account</a>
+                                    <a href="{% url 'register' %}">Create new account</a>
                                 </p>
                                 <p>
                                     <a href="#!">Forgot password?</a>
